refactored the ml.rnn to ml.nn and added tests; updated the models submodule
