+++ bin/SuperFluxNN
@@ -11,13 +11,13 @@ import glob
 import argparse
 
 from madmom import MODELS_PATH
-from madmom.processors import IOProcessor, io_arguments
+from madmom.processors import ParallelProcessor, IOProcessor, io_arguments
 from madmom.audio.signal import SignalProcessor, FramedSignalProcessor
 from madmom.audio.filters import FilterbankProcessor
 from madmom.audio.spectrogram import (LogarithmicSpectrogramProcessor,
                                       SpectrogramDifferenceProcessor,
                                       SuperFluxProcessor)
-from madmom.ml.rnn import RNNProcessor, average_predictions
+from madmom.ml.nn import NeuralNetwork, average_predictions
 from madmom.features import ActivationsProcessor
 from madmom.features.onsets import SpectralOnsetProcessor, PeakPickingProcessor
 
@@ -82,7 +82,7 @@ def main():
     args.fps = 100
     args.online = False
     args.onset_method = 'superflux'
-    args.nn_files = glob.glob("%s/onsets/2014/onsets_brnn_pp_[1-8].npz" %
+    args.nn_files = glob.glob("%s/onsets/2014/onsets_brnn_pp_[1-8].pkl" %
                               MODELS_PATH)
 
     # print arguments
@@ -106,9 +106,10 @@ def main():
         # save the Onset activations to file
         out_processor = ActivationsProcessor(mode='w', **vars(args))
     else:
-        # perform RNN processing, averaging and peak-picking
-        rnn = RNNProcessor(**vars(args))
+        # process everything with multiple RNNs and average the predictions
+        rnn = ParallelProcessor([NeuralNetwork.load(f) for f in args.nn_files])
         avg = average_predictions
+        # perform peak picking, i.e. select the local maxima
         pp = PeakPickingProcessor(**vars(args))
         from madmom.utils import write_events as writer
         # sequentially process them
